+++ app/models/auth-tokens/index.ts
@@ -1,7 +1,6 @@
 import createDebug from 'debug';
 
 export { AuthToken } from './AuthToken';
-export { SessionTokenV0 } from './SessionTokenV0';
 export { AppTokenV1 } from './AppTokenV1';
 export { SessionTokenV1 } from './SessionTokenV1';
 export { SessionTokenV1Store } from './SessionTokenV1Store';