+++ config/default.js
@@ -323,8 +323,7 @@ config.authSessions = {
   reissueGraceIntervalSec: 300,
   // Cleanup
   activeSessionTTLDays: 60,
-  // TODO The inactiveSessionTTLDays should be reduced to a few days when the V0 sessions become deprecated
-  inactiveSessionTTLDays: 60,
+  inactiveSessionTTLDays: 7,
   cleanupIntervalSec: 3600, // one hour
 };